+++ core/renderer/ui_wrapper/painting/harmony/painting_context_harmony.h
@@ -79,6 +79,8 @@ class PaintingContextHarmony : public PaintingCtxPlatformImpl {
   std::vector<float> getBoundingClientOrigin(int id) override;
   std::unique_ptr<pub::Value> GetTextInfo(const std::string& content,
                                           const pub::Value& info) override;
+  void StopExposure(const pub::Value& options) override;
+  void ResumeExposure() override;
   std::vector<float> getWindowSize(int id) override;
   std::vector<float> GetRectToWindow(int id) override;
   /**