+++ platform/harmony/lynx_harmony/src/main/cpp/ui/ui_owner.cc
@@ -611,10 +611,14 @@ void UIOwner::RemoveUIFromExposedMap(UIBase* ui, std::string unique_id) {
 void UIOwner::TriggerExposureCheck() { ui_observer_->TriggerExposureCheck(); }
 
 void UIOwner::StopExposure(const lepus::Value& options) {
-  ui_observer_->StopExposure(options);
+  auto task = [this, options]() { ui_observer_->StopExposure(options); };
+  context_->RunOnUIThread(std::move(task));
 }
 
-void UIOwner::ResumeExposure() { ui_observer_->ResumeExposure(); }
+void UIOwner::ResumeExposure() {
+  auto task = [this]() { ui_observer_->ResumeExposure(); };
+  context_->RunOnUIThread(std::move(task));
+}
 
 void UIOwner::SetObserverFrameRate(const lepus::Value& options) {
   ui_observer_->SetObserverFrameRate(options);